+++ packages/data/src/components/registry-provider/index.js
@@ -1,0 +1,15 @@
+/**
+ * WordPress dependencies
+ */
+import { createContext } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import defaultRegistry from '../../default-registry';
+
+const { Consumer, Provider } = createContext( defaultRegistry );
+
+export const RegistryConsumer = Consumer;
+
+export default Provider;
