+++ packages/data/src/default-registry.js
@@ -1,0 +1,3 @@
+import { createRegistry } from './registry';
+
+export default createRegistry();
