+++ packages/data/src/store/index.js
@@ -1,16 +1,12 @@
 /**
  * Internal dependencies
  */
-import { registerStore } from '../';
-
 import reducer from './reducer';
 import * as selectors from './selectors';
 import * as actions from './actions';
 
-export default function registerDataStore() {
-	registerStore( 'core/data', {
-		reducer,
-		actions,
-		selectors,
-	} );
-}
+export default {
+	reducer,
+	actions,
+	selectors,
+};
